CSCQVAIN-333: License validation fixes (#158)
* validate after openRecord

* fix license validation always showing error for free-form license urls

* add generic component used by oneOf and anyOf
