+++ src/widgets/anyOf.vue
@@ -1,28 +1,15 @@
-<!-- ADD_LICENSE_HEADER -->
-<template>
-	<div :id="property + '_anyOf'">
-		<!-- (anyof component) -->
-		<h2 v-bind:class="{ 'metainfo': !schema.title, 'missing': !schema.title }">{{ schema.title || "missing title" }}</h2>
-
-		<p v-if="schema.description">{{ schema.description }}</p>
-		anyOf: {{ schemaState }}
-
-		<div v-for="(sub, i) in schema['anyOf']" :key="sub">
-			{{ sub['type'] }} {{ i }}
-
-			<TabSelector :schema="sub" :path="newPath('anyOf/' + i)" :value="value" :parent="parent" :property="property" :tab="myTab" :activeTab="activeTab"></TabSelector>
-		</div>
-
-	</div>
-</template>
-
 <script>
-import vSchemaBase from './base.vue'
-
 export default {
-	extends: vSchemaBase,
-	name: 'schema-anyof',
+	name: 'SchemaAnyof',
 	description: "generic anyof",
-	schematype: '',
+	schematype: 'any',
+	functional: true,
+	render(createElement, { props, children }) {
+		const attrs = {
+			...props,
+			choiceProp: "anyOf",
+		}
+		return createElement('schema-choice', { attrs }, children)
+	},
 }
 </script>
