+++ src/widgets/mapping.js
@@ -10,6 +10,7 @@ export default {
 	'SchemaAllOf': require('./allOf.vue').default,
 	'SchemaOneOf': require('./oneOf.vue').default,
 	'SchemaEnum': require('./Enum.vue').default,
+	'SchemaChoice': require('./Choice.vue').default,
 	'reference-data': require('../components/ReferenceData.vue').default,
 	'i18n-string': require('@/components/i18nStrings.vue').default,
 	'i18n-textarea': require('@/components/i18nTextarea.vue').default,
